+++ R/analysis.R
@@ -59,11 +59,15 @@ svy_prop <- function(design, col, group = NULL, na_rm = T, stat_name = "prop", .
 
   to_return <- design |>
     srvyr::group_by(dplyr::across({{ group }}), dplyr::across({{ col }})) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_prop(...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_prop(...),
+      "n_unw" := srvyr::unweighted(srvyr::n())) |>
+    srvyr::mutate("{stat_name}_unw" := prop.table(.data$n_unw)) |>
     srvyr::ungroup()
 
   return(to_return)
 }
+
 
 
 #' @title Survey mean
@@ -87,7 +91,10 @@ svy_mean <- function(design, col, group = NULL, na_rm = T, stat_name = "mean", .
 
   to_return <- design |>
     srvyr::group_by(dplyr::across({{ group }})) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_mean({{ col }},...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_mean({{ col }},...),
+      "{stat_name}_unw" := srvyr::unweighted(mean({{ col }})),
+      "n_unw" := srvyr::unweighted(srvyr::n())) |>
     srvyr::ungroup()
 
   return(to_return)
@@ -116,7 +123,10 @@ svy_median <- function(design, col, group = NULL, na_rm = T, stat_name = "median
 
   to_return <- design |>
     srvyr::group_by(dplyr::across({{ group }})) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_median({{ col }}, ...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_median({{ col }}, ...),
+      "{stat_name}_unw" := srvyr::unweighted(stats::median({{ col }})),
+      "n_unw" := srvyr::unweighted(srvyr::n()))|>
     srvyr::ungroup()
 
   return(to_return)
@@ -146,7 +156,9 @@ svy_count_numeric <- function(design, col, group = NULL, na_rm = T, stat_name = 
   to_return <- design |>
     srvyr::mutate("{{ col }}" := as.character({{ col }})) |>
     srvyr::group_by(dplyr::across({{ group }}), dplyr::across({{ col }})) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_prop(...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_prop(...),
+      "n" := srvyr::unweighted(srvyr::n())) |>
     srvyr::ungroup()
 
   return(to_return)
@@ -179,12 +191,15 @@ svy_interact <- function(design, interact_cols, group = NULL, unnest_interaction
   to_return <- design |>
     srvyr::group_by(srvyr::across({{ group }}),
                     srvyr::interact(interaction = srvyr::across({{ interact_cols }}))) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_mean(...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_mean(...),
+      "n_unw" := srvyr::unweighted(srvyr::n())) |>
     dplyr::arrange(dplyr::desc(.data$prop)) |>
+    srvyr::mutate("{stat_name}_unw" := prop.table(.data$n_unw)) |>
     srvyr::ungroup()
 
   if (rlang::is_true(unnest_interaction)){
-    to_return <- to_return |> tidyr::unnest(.data$interaction)
+    to_return <- to_return |> tidyr::unnest("interaction")
   }
 
   return(to_return)
@@ -215,7 +230,10 @@ svy_ratio <- function(design, num, denom, group = NULL, na_rm = T, stat_name = "
 
   to_return <- design |>
     srvyr::group_by(dplyr::across({{ group }})) |>
-    srvyr::summarize("{stat_name}" := srvyr::survey_ratio({{ num }}, {{ denom }},...)) |>
+    srvyr::summarize(
+      "{stat_name}" := srvyr::survey_ratio({{ num }}, {{ denom }},...),
+      "{stat_name}_unw" := srvyr::unweighted(sum({{ num }}) / sum({{ denom }})),
+      "n" := srvyr::unweighted(srvyr::n())) |>
     srvyr::ungroup()
 
   return(to_return)
@@ -605,7 +623,7 @@ make_analysis_from_dap <- function(
 
   mapped <- purrr::pmap(
     dap |>
-      dplyr::select(.data$question_name, .data$analysis, .data$none_label, .data$group, .data$level, .data$na_rm, .data$vartype),
+      dplyr::select("question_name", "analysis", "none_label", "group", "level", "na_rm", "vartype"),
     function(question_name, analysis, none_label, group, level, na_rm, vartype){
 
       if (na_rm == "TRUE") {na_rm_lgl <- TRUE} else { na_rm_lgl <- FALSE}
@@ -622,6 +640,11 @@ make_analysis_from_dap <- function(
     purrr::set_names(dap$id_analysis)
 
   if (bind) {
+
+    if(sum(is.na(dap[["id_analysis"]])) > 0) rlang::abort("There are missing values in column 'id_analysis'. In order to bind results, all values in 'id_analysis' must exist and be unique.")
+
+    if(length(dap[["id_analysis"]]) > length(unique(dap[["id_analysis"]]))) rlang::abort("There are duplicated values in column 'id_analysis'. In order to bind results, all values in 'id_analysis' must exist and be unique.")
+
     mapped <- mapped |>
       purrr::map(\(x) x |> dplyr::mutate(dplyr::across(dplyr::starts_with("choices"), as.character))) |>
       dplyr::bind_rows(.id = "id_analysis") |>